requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.2
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
torch==2.1.2
werkzeug==3.0.1
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2
chardet==5.2.0

